requirements.txt
streamlit
soundfile
streamlit-webrtc
openai-whisper
SpeechRecognition
pydub
twilio
langchain
openai
python-dotenv